bot/bot_requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html

numpy==1.19.5
Pillow==8.2.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
APScheduler==3.6.3
certifi==2020.12.5
pkg-resources==0.0.0
python-telegram-bot==13.4.1
pytz==2021.1
six==1.15.0
tornado==6.1
tzlocal==2.1